% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nolap.R
\name{nolap}
\alias{nolap}
\title{Arrange data points so that they do not overlap in a scatter plot.}
\usage{
nolap(
  x,
  y,
  xdiv = 70,
  ydiv = 50,
  xlim = range(x),
  ylim = range(y),
  seed = NULL,
  nseed = 10
)
}
\arguments{
\item{x}{vector of x-coordinates}

\item{y}{vector of y-coordinates}

\item{xdiv}{number of x-axis divisions}

\item{ydiv}{number of y-axis divisions}

\item{xlim}{range of x-axis to be divided}

\item{ylim}{range of y-axis to be divided}

\item{seed}{initial value of the random number to be specified}

\item{nseed}{number of trials when initial random number is not specified}
}
\value{
xleft: a vector of left x positions.

ybottom: a vector of bottom y positions.

xright: a vector of right x positions.

ytop: a vector of top y positions.

x: x-coordinate of the center of the rectangular region

y: y-coordinate of the center of the rectangular region

xlims: x-axis dividing interval

ylims: y-axis dividing interval

ssd: sum of squared total distance moved

seed: The initial value of the specified random number or the best initial value obtained by trial
}
\description{
\code{nolap} This package aims to optimize arrangement by minimizing the movement of data points so that they do not overlap.
}
\examples{
x <- cars$speed
y <- cars$dist
res <- nolap(x=x,y=y,xdiv=10,ydiv=10,nseed=10)
print(res$seed)
plot(res$x,res$y,col=2,xlim=range(res$xlims),ylim=range(res$ylims))
abline(h=res$ylims,col=8)
abline(v=res$xlims,col=8)
# check movements
plot(x,y,xlim=range(res$xlims),ylim=range(res$ylims),type="n")
text(x,y,1:50)
abline(h=res$ylims,col=8)
abline(v=res$xlims,col=8)
arrows(x,y,res$x,res$y,length=0.1,col=ifelse(res$is.moved,2,4))
legend("topleft",legend=c("moved","stay"),fill=c(2,4))
}
